third_party/WebKit/LayoutTests/http/tests/media/media-document.html

<!DOCTYPE html>
<title>Tests that a media document can open a media URL with an ambiguous extension.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../../media-resources/media-file.js"></script>
<body>
<script>
async_test(function(t) {
    var movie = 'test.ogv';
    var type = mimeTypeForExtension(movie.split('.').pop());
    var iframe = document.createElement('iframe');

    iframe.onload = t.step_func_done(function() {
        assert_equals(iframe.contentDocument.querySelector('video').error, null);
    });

    iframe.src = 'resources/video-check-useragent.php?name=' + movie + '&type=' + type;
    document.body.appendChild(iframe);
});
</script>
